[Refactor] [10/N] to simplify the vLLM openai completion serving architecture
