+++ vllm/entrypoints/sagemaker/routes.py
@@ -12,22 +12,26 @@ from fastapi.responses import JSONResponse, Response
 
 from vllm.entrypoints.openai.api_server import (
     base,
-    chat,
-    completion,
-    create_completion,
-    validate_json_request,
 )
 from vllm.entrypoints.openai.chat_completion.api_router import (
+    chat,
     create_chat_completion,
 )
 from vllm.entrypoints.openai.chat_completion.protocol import (
     ChatCompletionRequest,
 )
-from vllm.entrypoints.openai.engine.protocol import (
+from vllm.entrypoints.openai.completion.api_router import (
+    completion,
+    create_completion,
+)
+from vllm.entrypoints.openai.completion.protocol import (
     CompletionRequest,
+)
+from vllm.entrypoints.openai.engine.protocol import (
     ErrorResponse,
 )
 from vllm.entrypoints.openai.engine.serving import OpenAIServing
+from vllm.entrypoints.openai.utils import validate_json_request
 from vllm.entrypoints.pooling.classify.api_router import classify, create_classify
 from vllm.entrypoints.pooling.classify.protocol import ClassificationRequest
 from vllm.entrypoints.pooling.embed.api_router import create_embedding, embedding
